backend/requirements.txt
# Django & DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Task Queue
celery==5.3.4
redis==5.0.1

# Network Integration
librouteros==3.1.0
pysnmp==4.4.12

# Payment Integrations
stripe==7.8.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
Pillow==10.1.0

# Additional dependencies for API improvements
django-redis==5.4.0
psutil==5.9.6
python-dateutil==2.8.2
